Login Page
* Refactored Login Page to use react hook form to handle form state
* Converted page to class component and split up form to it's own
  component
* Added javascript stylings
* Removed warnings
