+++ views/pages/ConfigPage.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { useOutletContext, Link } from 'react-router-dom'
