+++ views/styles/index.js
@@ -4,6 +4,7 @@ import { configStyles } from './config_styles'
 import { graphPageStyles } from './graph_page_styles'
 import { studyDetailStyles } from './study_details'
 import { mainLayoutStyles } from './main_layout_styles'
+import { loginStyles } from './login_page_styles'
 import { shareFormStyles } from './share_form_styles'
 
 const drawerWidth = 200
@@ -16,6 +17,7 @@ export const styles = (theme) => ({
   ...mainLayoutStyles(theme),
   ...accountPageStyles(theme),
   ...shareFormStyles(theme),
+  ...loginStyles(theme),
   selectedFontWeight: {
     fontWeight: 500,
   },
